% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationApp.R
\name{locationApp}
\alias{locationApp}
\title{Get the presslog for Ames PD for an interval}
\usage{
locationApp(...)
}
\value{
shiny app with location of calls
}
\description{
Get the presslog for Ames PD for an interval
}
\examples{
\dontrun{
locationApp()
}

}
\author{
Tyler Wiederich
}
